Add a context menu to run PDF OCR.
Adds a menu item that will be only present in the context menu of the
page when a PDF is opened and Screen Reader is enabled.
It triggers running Screen AI visual annotation to extract PDF content.

This CL is based on crrev.com/c/3718362 by nektar@.

Bug: 1278249
Change-Id: Ia2119f9c631084731088d2c3b435235047cf19aa
Ax-Relnotes: n/a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3723258
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/main@{#1019476}