ime: Add test only API for checking if an IME is ready to accept keys.
Tast tests need a way to tell whether an IME is ready or not before
sending key events. If it sends it too early, then the IME will simply
type out latin characters, instead of composing them into the target
script.

Add a new autotestPrivate API for this. We need to plumb all the way
into IME Service (as that's where the IME loading happens), so we also
add a new Mojo API. The new method will be implemented in a separate
internal repo in a followup CL.

For now, this API only works for "System PK" IMEs. It will always
return 'false' for other IMEs.

Bug: b/195374149
Change-Id: I9b08ba073e8715e5300a18433eaa958390c93cb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3689105
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Reviewed-by: John Palmer <jopalmer@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019505}