[Signin] Record Oauth2 response regardless of the response code
This CL moves from `Gaia.BadRequestTypeForOAuth2AccessToken` histogram
that is only recorded if Oauth2 response code is 400 (bad request) to a
more generic histogram recorded for any response.

Note: Some of the buckets like `OAUTH2_ACCESS_ERROR_INVALID_CLIENT` were
never recorded as the server sends a response code 401 (Unauthorized)
for this error not 400.

`OAuth2ErrorCodes` will be used in a future CL to fire the right
`GoogleServiceAuthError` as the response code is not sufficient.

Bug: 1336627
Change-Id: If5b813479a9828dd6a45ce7a1295cebb59c3cf0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3710093
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Weilun Shi <sweilun@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019574}