[GPUChannel] Mark accessor as test-only
GPUChannel exposes ImageDecodeAcceleratorStub via an accessor. On
inspection, this accessor is used only by tests. Mark it as ForTesting
to ensure that this stays the case.

In addition to just being good practice, this helps concretely as
ImageDecodeAcceleratorStub heavily uses GLImage, which we are trying
to eliminate, and thus locking down its accessibility is a good thing.

Bug: 1310018
Change-Id: I543148d1ac2d127a79245794b2997f416afde7ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3737744
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Auto-Submit: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019603}