[FLEDGE] Block renderer from reading auction-only signals.
Only the browser process should be able to read signals intended for
FLEDGE auction participants -- the renderer shouldn't be able to read
these signals.

Introduce a new response header, X-FLEDGE-Auction-Only. If this header
is present on a response, only the browser process will be allowed to
read the response.

X-FLEDGE-Auction-Only is created instead of augmenting the semantics
of the existing X-Allow-FLEDGE header, which is a header required for
all network responses used in FLEDGE auctions -- this is because several
network requests made with X-Allow-FLEDGE use the network factory of the
frame that called navigator.runAdAuction(), so these requests are
treated as if they are issued by the renderer.

Bug: 1186444,1320908
Change-Id: I2c3b0499dee8ca90a5ee2b2b8f07470f9e20226a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3653083
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Commit-Queue: Caleb Raitto <caraitto@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019604}