[cc] Subsample per-draw metrics
Some metrics are recorded each time we draw. This is unnecessary, we
draw frames often enough to not need all these data points, and metrics
recording is not completely free. Furthermore, this code actually
acquires a lock at each call, which might cause issues (even though the
lock is not contended).

As a result, only record the metrics 1% of the time. Using

the default RNG makes a system call at every sample.

base: :InsecureRandomGenerator as this is only for metrics recording, and
Change-Id: I9354fe26a5f8d821e7f0ca921157266fafa8650e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3688092
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019643}