[Lacros] Support ProfilePickerOnStartupAvailability policy
The profile picker has been ported to Lacros. This CL enables the
associated policy ProfilePickerOnStartupAvailability on Lacros as well.

There is no test in this CL, because the behavior is already tested in:
ProfilePickerTest.ShouldShowAtLaunch_MultipleProfiles_TwoActive_Disabled
This test is already enabled on Lacros, and uses the pref
prefs::kBrowserShowProfilePickerOnStartup to control the behavior.

Change-Id: I9fdf202f9c2266ddc41ce2e14c37052c72535a3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3732398
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019652}