Fill in functionality that initiates retrying a download
(CL 2 of 3)
This CL kicks off retrying the download. It uses a new download source
RETRY_FROM_BUBBLE to distinguish this from the RETRY option that is used
in chrome://downloads as well as on Android. The source URL used is the
last one of the original redirect chain to restrict down to a previously vetted URL since there is no associated RenderFrameHost to use for the DownloadUrlParameters.

Bug: 1331584
Change-Id: I0fc636f4f2ba996adee8ecf7251bcc3d22bdee31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3730143
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019669}