Enable CSPP killswitch on Mac by default
The Finch team requires that when a killswitch config is landed, we
follow up by merging the default behavior to the affected branches.
Since we're already merging the creation of the killswitch to M104, it
doesn't make sense to apply the Finch config. Instead, this CL activates
the killswitch on Mac, and we can merge both this CL and
https://crrev.com/c/3732758 to M104.

Bug: 1326455
Change-Id: Id7535cf113cadfe862ba3ac181cca8c9a5bcf2c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3736982
Reviewed-by: Rohit Bhatia <bhatiarohit@google.com>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019711}