[Extensions] Fix potential UAF in UpdateDataProvider.
UpdateDataProvider post a task that can be run after its BrowserContext is destroyed, leading to user-after-free. This CL instead binds the UpdateDataProvider instance to the callback and uses a member function instead of a static function. This ensures the the UpdateDataProvider instance will be alive when the callback is executed.

Bug: 1323449
Change-Id: If617676d2e0e6a31a7936d88c75dd221c69b1057
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3688185
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: David Bertoni <dbertoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019720}