[ios] Add timeout for slow HTTPS loads in Omnibox HTTPS upgrades
This CL implements the slow-load timer for Omnibox HTTPS Upgrades
feature. The feature observes navigations upgraded by the omnibox to
HTTPS and falls back to an HTTP navigation if the HTTPS page fails to
load or doesn't return a response within the allotted time (3 seconds).

The timer is started when the HTTPS upgrade is detected. When the
timer times out, the tab helper stops the current page load and
initiates a fallback navigation to the HTTP version of the same URL.

This CL completes the main implementation of the Omnibox HTTPS
Upgrades feature.

Bug: 1195440
Change-Id: I5c4da842c791aef61c2ad8eb297a6eb9064e7a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3727482
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019723}