Report if archive analysis times out or the archive is too large
These are the top two reasons for archives to not be parsed, and we
currently report "INVALID_" for both of them. The server-side analysis
could check the size field to determine if the archive is too large, so
that field is optional. But they currently have no way of knowing if
the analysis timed out.

Bug: 1314481
Change-Id: I54e1b3574f89df6c6c099c5ba3d67e96423c48f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708185
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019728}