[NTP] Use photos homepage for test data urls
When #ntp-photos-module is set to any of the fake data options, clicking on a memory will send the user to a 404 page. Since the data is faked, we do not have a real memory to send the user to. Instead, a check has been added when the url is created to send the user to the photos homepage for fake memories.

Bug: 1334986
Change-Id: I0ad676235e70592aa4ee683debd01e9eac96933b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3718807
Commit-Queue: Riley Tatum <rtatum@google.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019750}