browser: tolerate RenderFrameHost death with link generation menu open
If:
1. The context menu is open with the link generation item in it, and
2. The RenderFrameHost crashes, and
3. The link generation menu's timeout response arrives after the crash,

then the browser, too, will crash trying to use the dead
RenderFrameHost. To avoid that, retain the RenderFrameHost only as an ID
and look it up each time it is needed. Also, avoid trying to use Mojo
remotes we got from that RFH if the RFH is dead, since the remotes will
be unbound at that point as well.

Fixed: 1329794
Change-Id: I57a3471f8e7fdd84373006eacc30f2da5b3e3f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3705832
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: Jeffrey Cohen <jeffreycohen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019761}