[APC Desktop] Introduce callback to signal success of a run.
This CL changes to signature of ApcClient::Start() to remove the
return value and add a callback that signals whether the run ended
successfully or not.
Previously, the return value would indicate only whether a run was
started successfully and not whether would complete successfully.
However, the latter is more successful in deciding whether to adjust
an entry in the list of leaked credentials.

Bug: 1340073
Change-Id: I999ccb7a1a22c3080bb00fdb11c3f7fb0a1bf2b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735050
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1019798}