[webauthn] Add "Sign in with passkeys…" button
Add a "Sign in with passkeys…" button to the autofill / webauthn
conditional mediation prompt that starts a regular WebAuthn request.
Plugged in authenticators are only dispatched to after the user clicks
this button.

This also tweaks how restarting a request when a Conditional UI
request fails or is cancelled works. We should always StartOver(),
even if the request is not in a finished state yet. This resets the
handlers which cancels ongoing requests while letting the user restart
using the UI. This was made evident playing with the new button.

Fixed: 1329957
Change-Id: I737fd1715ee77a869dd71b9db1a4c93aab3a9425
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3708247
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019822}