Added new view for optional toolbar button
ToolbarPhone has an optional button that shows up between the omnibox
and the tab switcher button. This button is mainly used to surface
features on a user-by-user basis (a.k.a. Magic Toolbar), it contains a
show/hide animation.

The contextual page action (CPA) project aims to surface features based
on the current website. With the limited real estate on the toolbar we
decided to replace the Magic Toolbar action with a contextual action
when loading compatible websites. We are aiming to experiment with two
main variations, one that swaps the icons in place and another one that
exposes the new action on an action chip which then colapses back into
a regular icon.

This CL adds a new view OptionalButtonView which shows a toolbar
button whose icon can be updated with one of two transitions: swapping
or expanding as an action chip. A follow-up CL (https://chromium-review.googlesource.com/c/chromium/src/+/3698772) creates MVC classes
and a final CL (https://chromium-review.googlesource.com/c/chromium/src/+/3697050/) integrates this
component into ToolbarPhone.

Figma spec:
https://www.figma.com/proto/xDx5P5tjLf8CXD70BY6wqP/Contextual-Actions?page-id=334%3A6697&node-id=334%3A8441&viewport=307%2C48%2C0.15&scaling=min-zoom&starting-point-node-id=334%3A8308&show-proto-sidebar=1

PRD:
https://docs.google.com/document/d/10ApoEZwkUS5LsqDw1YjGtDOWzfXrukpNeuocJQI6xkI/edit?usp=sharing

Animation technical details:
https://docs.google.com/document/d/16IiK39BdCw3iCbQijT2BWE99rZ0r1ueE3TDtNLC4uRc/edit?usp=sharing

Original CL in experimental branch:
https://chromium-review.googlesource.com/c/experimental/chromium/src/+/3578413


Bug: 1301633
Change-Id: Ic9967defb7d7fa5eaf566190915c8e7db410f539
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3700608
Reviewed-by: Matthew Jones <mdjones@chromium.org>
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019845}