Experiment to Measure client prefetching on selection change
This is a precursor to a more robust solution involving specific events.
This CL will add a prefetch for Desktop omnibox interactions where the
user either clicks or uses the down arrow to change selection. This is
an approximation of the behavior we want and will help inform the impact
vs precision we can achieve in this domain. This replaces the previous
experiment that prefetched at navigation. This is more scoped to not the
first suggestion and before the actual navigation start, so these are
less likely to be used than those at navigation start.

Bug: 1314128, 1318154, 1317890
Change-Id: I67eb9b45e2fc83d9e8261346ce2fbe590a267fb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3712127
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1019880}