Remove FLoC toggle preference
As the feature has been removed from the codebase and the generated
pref was already disabling the user control, this CL aims to cleanup
dead code for the FLoC preference while maintaining the UI as
disabled until Privacy Sandbox Settings 3 is launched.
A follow-up is to be expected to remove all the remaining FLoC code
along with its UI elements.

Bug: 1299720
Change-Id: I882a9bcecddaa3907bd804ab7bbaf8707b124f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3726359
Commit-Queue: Mariam Ali <alimariam@google.com>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020002}