CCA: Disable right-click behavior on preview element
There is a known issue that when mirror button is on, the video control
will also be flipped. But actually we should not show video controls to
users for preview element. Therefore, this CL simply disables right-click
behavior on preview element and should be able to solve this issue.

Bug: b:178660469
Test: Manually
Change-Id: I39be60df21e1e00564987183458880d3425ef37b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3740045
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020007}