[iOS] Re-enabling EG test for Enhanced Safe Browsing(ESB)
The EG test was disabled due to the default activation point of
TableViewInfoButtonItem's to be on the "i" button instead of the center
of the row/cell. Since that has been resolved in the ESB menu, we can
re-enable this test. Additionally, improvements were made to the
EG test to make sure it worked successfully and to ensure that
preference values were set properly in the EG test's set up and tear
down.

Bug: 1333625, 1307789, 1329733
Change-Id: I005c1c33e52c7eb9a56966df817279886e74fd27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3735159
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020098}