dPWA: Fix testing logic and dependencies for external pref migration
This CL performs the following tasks:
1. Fixes test logic w.r.t the external pref migration release.
2. Moves pref and DB writes to web_app_test_utils to have a single
surface on where all data can be written to keep stuff in sync.
3. Fixes test failures from crrev.com/3726413 when the release flag
is enabled.

This does not modify any existing system behavior or change the
location of how the writes happen on production (non-test code), mainly
to prevent double opening of the DB when the data is written from the
WebAppInstallFinalizer.

Bug: 1029410
Change-Id: I6a997b73fd9dfeb52c242fd7d2fe6e8ba9a6d693
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3721892
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Christopher Lam <calamity@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020099}