cros: set highlight border above browser window
Current highlight border of browser window is too thick which looks
not good. To avoid this, we want the inner border overlap the web
content and outer border outside the content.

To make inner border overlap the content, we should paint the border
above whole browser window. This cl adds a new class
`HighlightBorderOverlay` which creates a nine patch layer painted with
a highlight border and overlay on top of the widget.

When the outer border beyond the window contents in the dark mode,
we should decrease the opacity of the outer border to 6%. Therefore,
we set a new border type kHighlightBorder3 to achieve this goal.

The final effects are shown in the bug comment.

Bug: 1334719
Change-Id: I1ec7f7eebc5e8b807247872dab93bb8a79c373aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3698761
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Min Chen <minch@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020104}