webhid: Expose WebHID on service workers under a blink feature flag
This is part of exposing WebHID to extension service workers changes.
This change exposes WebHID on service workers under a blink feature flag
WebHIDOnServiceWorkers. This flag currently in this patch is a
placeholder and the following patches of this change series will enable
it according to a chrome feature flag in the extension environment.

Bug: 1303195
Change-Id: I2c85fd3fed71ea2723bda8226550050ac094990c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3680193
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: Jack Hsieh <chengweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020109}