StableVideoDecoderService: Implement OnWaiting().
This CL implements the StableVideoDecoderService::OnWaiting() method
which receives a notification from the underlying mojom::VideoDecoder
when it needs to tell the client that it's waiting for something (e.g.,
a decryption key). StableVideoDecoderService::OnWaiting() only needs to
forward this notification to the stable::mojom::VideoDecoderClient.

The unit tests are expanded to cover the new code.

Bug: b:195769334
Test: ./media_unittests --gtest_filter="StableVideoDecoderServiceTest.*" in amd64-generic VM.
Change-Id: Ibfd1eb44a3abca950e145bfa5d3755025fa75ed5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3683913
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Pilar Molina Lopez <pmolinalopez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020123}