Fix percentage of tabs used metric
Reset the initial tabs value to 0 when application stops. Another theory is that initialTabsCount is not being set to the correct number in some cases. This could happen if the tab model state is not initialized. This also adds a listener for the tab model state initialization to get the initial count from there as a backup.

Bug: 1341083
Change-Id: I7ad3d841186cd4188c4b97fad289619d2538ec49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3739043
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Commit-Queue: Gaurav Jayasawal <gauravjj@google.com>
Cr-Commit-Position: refs/heads/main@{#1020132}