Fix similar if-statements in autofill external delegate
This CL fixes having two of the same if-statements in autofill external delegate.

It also ensures that the Autocomplete Suggestion Accepted Index is only logged when autocomplete suggestions are accepted, as the current logic logs this metric for both autocomplete suggestions and merchant promo
code suggestions.

Bug: 1190334
Change-Id: Ib3f16f5b16767153717127a4e6c36cf8e548e74a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3712231
Reviewed-by: Jared Saul <jsaul@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1020151}