Clear persisted SCT reports when browsing history is cleared
This adds a call to SCTAuditingHandler::ClearPendingReports() to
NetworkContext::ClearNetworkingHistoryBetween() and adds support for the
BarrierClosure so that we ensure the ImportantFileWriter commits the
cleared state to disk. This follows the general pattern used for
clearing the TransportSecurityState (which is also cleared when history
is cleared).

Bug: 1144205
Change-Id: Iff6672418d02a620e86207692c6578cb19579769
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3408677
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Auto-Submit: Chris Thompson <cthomp@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: John Abd-El-Malek <jam@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020180}