[Nearby] Fix cancel button regression
Fixes a small behavior change with the `close` event being fired. This
change makes the code that fires the event set the `event.detail` to an
empty object instead of sending null, which caused the handler to break.

A unit test was also modified to catch this event properly.

Test=Unit tests, manually verified

Fixed: b/237703787, 1341306
Change-Id: I837b89dbbe79e24eb875969da6bff6e5057d5a33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3739064
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1020187}